automl/requirement.txt
matplotlib==2.2.3
numpy==1.15.4
pandas==0.23.4
google-cloud-storage
google-cloud-automl_v1beta1
grpcio==1.17.1
Pillow==5.3.0
protobuf==3.6.1
scipy==1.2.0
